CMSIS-DSP: Removed float promotion issues Corrected some additional warnings when compiling with AC5 compiler. Updated version numbers,

+++ CMSIS/DSP/Source/BasicMathFunctions/arm_abs_f16.c
@@ -156,13 +156,13 @@ void arm_abs_f16(
     /* C = |A| */
 
     /* Calculate absolute and store result in destination buffer. */
-    *pDst++ = fabsf(*pSrc++);
+    *pDst++ = (_Float16)fabsf((float32_t)*pSrc++);
 
-    *pDst++ = fabsf(*pSrc++);
+    *pDst++ = (_Float16)fabsf((float32_t)*pSrc++);
 
-    *pDst++ = fabsf(*pSrc++);
+    *pDst++ = (_Float16)fabsf((float32_t)*pSrc++);
 
-    *pDst++ = fabsf(*pSrc++);
+    *pDst++ = (_Float16)fabsf((float32_t)*pSrc++);
 
     /* Decrement loop counter */
     blkCnt--;
@@ -184,7 +184,7 @@ void arm_abs_f16(
     /* C = |A| */
 
     /* Calculate absolute and store result in destination buffer. */
-    *pDst++ = fabsf(*pSrc++);
+    *pDst++ = (_Float16)fabsf((float32_t)*pSrc++);
 
     /* Decrement loop counter */
     blkCnt--;
